chore: 관심사 및 파일 분리
